+++ src/Gargantext/Database/Admin/Types/Hyperdata/Frame.hs
@@ -58,9 +58,9 @@ instance FromField HyperdataFrame
   where
     fromField = fromField'
 
-instance QueryRunnerColumnDefault PGJsonb HyperdataFrame
+instance DefaultFromField PGJsonb HyperdataFrame
   where
-    queryRunnerColumnDefault = fieldQueryRunnerColumn
+    defaultFromField = fieldQueryRunnerColumn
 
 instance ToSchema HyperdataFrame where
   declareNamedSchema proxy =