+++ src/Gargantext/Database/Query/Table/Node/Select.hs
@@ -31,8 +31,8 @@ selectNodesWithUsername nt u = runOpaQuery (q u)
   where
     q u' = proc () -> do
       (n,usrs) <- join' -< ()
-      restrict -< user_username usrs .== (toNullable $ pgStrictText u')
-      restrict -< _node_typename n .== (pgInt4 $ toDBid nt)
+      restrict -< user_username usrs .== (toNullable $ sqlStrictText u')
+      restrict -< _node_typename n .== (sqlInt4 $ toDBid nt)
       returnA  -< _node_id n
 
     join' :: Query (NodeRead, UserReadNull)