+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 
+## Version 0.0.4.1
+* Refact/code design better syntax for DataType fields
+
+## Version 0.0.4
+* Fix the search in Title and abstracts.
+* [UPGRADE] execute devops/postgres/upgrade/0.0.4.sql to your database to upgrade it
+
 ## Version 0.0.3.9.1
 * Graph Update fix
 * Document view: full text removed